Storage class improvements
readSave() and writeSave() are now the only functions in storage. buildTaskFromSave() is now under Task class instead (to better follow OOP standards).
TaskList also now supports its own toSave() function so as to not violate OOP standards.
