+++ src/main/java/cluck/storage/Storage.java
@@ -1,59 +1,55 @@
 package cluck.storage;
 
 import cluck.taskList.TaskList;
-import cluck.tasks.Deadline;
-import cluck.tasks.Event;
 import cluck.tasks.Task;
-import cluck.tasks.ToDo;
 
 import java.io.File;
 import java.io.FileNotFoundException;
-import java.util.Arrays;
+import java.io.FileWriter;
+import java.io.IOException;
 import java.util.Scanner;
 
 public class Storage {
+    private File saveFile;
 
-    public static Task buildTaskFromSave(String savedTask) {
-        String[] savedTaskFields = savedTask.split("\\|");
-        boolean isMarked;
-
-        if (savedTaskFields[1].equals("1")) {
-            isMarked = true;
-        } else if (savedTaskFields[1].equals("0")) {
-            isMarked = false;
-        } else {
-            System.out.println("Corrupted data found, skipping corrupted data.");
-            return null;
-        }
-
-        switch (savedTaskFields[0]) {
-        case "E":
-            return new Event(isMarked, savedTaskFields[2], savedTaskFields[3], savedTaskFields[4]);
-
-        case "D":
-            return new Deadline(isMarked, savedTaskFields[2], savedTaskFields[3]);
-
-        case "T":
-            return new ToDo(isMarked, savedTaskFields[2]);
-
-        default:
-            System.out.println("Corrupted data found, skipping corrupted data.");
-            return null;
-        }
+    public Storage(String filePath) {
+        saveFile = new File(filePath);
     }
 
-    public TaskList readSave(File savedFile) {
+    public TaskList readSave() {
+        if (saveFile.mkdirs()) {
+            return new TaskList();
+        }
         try {
             TaskList taskList = new TaskList();
             Task currTask;
-            Scanner savedFileScanner = new Scanner(savedFile);
+            Scanner savedFileScanner = new Scanner(saveFile);
             while (savedFileScanner.hasNextLine()) {
-                currTask = buildTaskFromSave(savedFileScanner.nextLine());
+                currTask = Task.buildTaskFromSave(savedFileScanner.nextLine());
                 taskList.addTask(currTask);
             }
             return taskList;
         } catch (FileNotFoundException e) {
             throw new RuntimeException(e);
+        }
+    }
+
+    /**
+     * Saves the current list of tasks into a txt file.
+     * This will overwrite previous saves.
+     * There should be no missing directory error since readSave()
+     * will create the save directory if it does not exist.
+     *
+     * @param taskList list of task to be saved
+     */
+    public void writeSave(TaskList taskList) {
+        try {
+            FileWriter writer = new FileWriter(saveFile);
+            writer.write(taskList.toSave());
+            writer.close();
+        } catch (IOException e) {
+            System.out.println("Buh oh! An error occurred!!");
+            e.printStackTrace();
         }
     }
 }
